requirements.txt
click
qiskit-aer
qiskit
flask
dask
typer
blinker
numpy
pygame
pycryptodome
